decorator: Extract validate_remote_server.
